Move header conversion to convert module
